+++ src/layouts/builder.vue
@@ -1,6 +1,5 @@
 <template>
   <q-layout view="hHr LpR lFf"> <!-- Be sure to play with the Layout demo on docs -->
-
     <!-- (Optional) The Header -->
     <q-layout-header reveal>
       <q-toolbar>